create sidebar component
